Move notifcations from indivual files in to a single header
