+++ libraries/plugins/apis/account_history_api/include/steem/plugins/account_history_api/account_history_api.hpp
@@ -2,7 +2,6 @@
 #include <steem/plugins/json_rpc/utility.hpp>
 
 #include <steem/chain/history_object.hpp>
-#include <steem/chain/operation_notification.hpp>
 
 #include <steem/protocol/types.hpp>
