+++ libraries/plugins/block_log_info/block_log_info_plugin.cpp
@@ -5,7 +5,6 @@
 #include <steem/chain/database.hpp>
 #include <steem/chain/global_property_object.hpp>
 #include <steem/chain/index.hpp>
-#include <steem/chain/operation_notification.hpp>
 
 #include <fstream>
 #include <iostream>
