+++ libraries/plugins/smt_test/smt_test_plugin.cpp
@@ -4,7 +4,6 @@
 #include <steem/chain/account_object.hpp>
 #include <steem/chain/database.hpp>
 #include <steem/chain/index.hpp>
-#include <steem/chain/operation_notification.hpp>
 
 #include <steem/protocol/smt_operations.hpp>
