+++ libraries/plugins/stats_export/stats_export_plugin.cpp
@@ -8,7 +8,6 @@
 #include <steem/chain/database.hpp>
 #include <steem/chain/global_property_object.hpp>
 #include <steem/chain/index.hpp>
-#include <steem/chain/operation_notification.hpp>
 
 #include <fstream>
 #include <iostream>
